Disable implicit casting by default and implement intrinsic functions for casting

+++ integration_tests/bindc_02.py
@@ -9,13 +9,13 @@ def f():
     yq: CPtr
     yptr1: Pointer[i16[:]]
     y: i16[2]
-    y[0] = 1
-    y[1] = 2
+    y[0] = i16(1)
+    y[1] = i16(2)
     yptr1 = pointer(y)
     print(pointer(y), yptr1)
     print(yptr1[0], yptr1[1])
-    assert yptr1[0] == 1
-    assert yptr1[1] == 2
+    assert yptr1[0] == i16(1)
+    assert yptr1[1] == i16(2)
 
     c_p_pointer(yq, yptr1)
