+++ tests/errors/test_bitwise_on_float.py
@@ -2,9 +2,9 @@ from ltypes import f32
 
 def f():
     f1: f32
-    f1 = 4.5
+    f1 = f32(4.5)
     f2: f32
-    f2 = 6.7
+    f2 = f32(6.7)
     print(f1 << f2)
